+++ test/vanilla/legacy/Expected/AcceptanceTests/BodyComplexPythonThreeOnly/bodycomplexpython3only/aio/operations/_polymorphism_operations.py
@@ -23,7 +23,7 @@ from azure.core.tracing.decorator_async import distributed_trace_async
 
 from ... import models as _models
 from ..._vendor import _convert_request
-from ...operations._polymorphism_operations import (
+from ...operations._polymorphism_operations_py3 import (
     build_get_complicated_request,
     build_get_composed_with_discriminator_request,
     build_get_composed_without_discriminator_request,