+++ test/vanilla/legacy/Expected/AcceptanceTests/BodyComplexPythonThreeOnly/bodycomplexpython3only/aio/operations/_primitive_operations.py
@@ -24,7 +24,7 @@ from azure.core.tracing.decorator_async import distributed_trace_async
 
 from ... import models as _models
 from ..._vendor import _convert_request
-from ...operations._primitive_operations import (
+from ...operations._primitive_operations_py3 import (
     build_get_bool_request,
     build_get_byte_request,
     build_get_date_request,